pyproject.toml
[project]
name = "schema-erd-viewer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "databricks-sdk>=0.97.0",
    "databricks-sql-connector>=4.2.5",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "uvicorn>=0.41.0",
]
